[reputation] add sled store
