+++ crates/icn-reputation/src/lib.rs
@@ -4,8 +4,12 @@ use icn_common::Did;
 #[cfg(test)]
 use icn_identity::ExecutionReceipt;
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::sync::Mutex;
+
+#[cfg(feature = "persist-sled")]
+pub mod sled_store;
+#[cfg(feature = "persist-sled")]
+pub use sled_store::SledReputationStore;
 
 /// Store for retrieving and updating executor reputation scores.
 pub trait ReputationStore: Send + Sync {
@@ -51,63 +55,11 @@ impl ReputationStore for InMemoryReputationStore {
     }
 }
 
-/// Persistent sled-backed reputation store.
-#[cfg(feature = "persist-sled")]
-pub struct SledReputationStore {
-    tree: sled::Tree,
-}
-
-#[cfg(feature = "persist-sled")]
-impl SledReputationStore {
-    /// Opens or creates a sled-backed store at the given path.
-    pub fn new(path: PathBuf) -> Result<Self, icn_common::CommonError> {
-        let db = sled::open(path).map_err(|e| {
-            icn_common::CommonError::DatabaseError(format!("Failed to open sled DB: {e}"))
-        })?;
-        let tree = db.open_tree("reputation_v1").map_err(|e| {
-            icn_common::CommonError::DatabaseError(format!("Failed to open tree: {e}"))
-        })?;
-        Ok(Self { tree })
-    }
-
-    fn read_score(&self, did: &Did) -> u64 {
-        if let Ok(Some(bytes)) = self.tree.get(did.to_string()) {
-            bincode::deserialize(&bytes).unwrap_or(0)
-        } else {
-            0
-        }
-    }
-
-    fn write_score(&self, did: &Did, score: u64) {
-        if let Ok(encoded) = bincode::serialize(&score) {
-            let _ = self.tree.insert(did.to_string(), encoded);
-            let _ = self.tree.flush();
-        }
-    }
-}
-
-#[cfg(feature = "persist-sled")]
-impl ReputationStore for SledReputationStore {
-    fn get_reputation(&self, did: &Did) -> u64 {
-        self.read_score(did)
-    }
-
-    fn record_execution(&self, executor: &Did, success: bool, cpu_ms: u64) {
-        let current = self.read_score(executor);
-        let base: i64 = if success { 1 } else { -1 };
-        let delta: i64 = base + (cpu_ms / 1000) as i64;
-        let updated = (current as i64) + delta;
-        let new_score = if updated < 0 { 0 } else { updated as u64 };
-        self.write_score(executor, new_score);
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use icn_identity::{did_key_from_verifying_key, generate_ed25519_keypair};
     use std::str::FromStr;
-    use tempfile::tempdir;
 
     #[test]
     fn reputation_updates() {
@@ -125,28 +77,5 @@ mod tests {
         };
         store.record_execution(&receipt.executor_did, receipt.success, receipt.cpu_ms);
         assert_eq!(store.get_reputation(&did), 1);
-    }
-
-    #[cfg(feature = "persist-sled")]
-    #[test]
-    fn sled_store_persists_scores() {
-        let dir = tempdir().unwrap();
-        let path = dir.path().join("rep.sled");
-        let store = SledReputationStore::new(path.clone()).unwrap();
-        let (_sk, vk) = generate_ed25519_keypair();
-        let did = Did::from_str(&did_key_from_verifying_key(&vk)).unwrap();
-        let receipt = ExecutionReceipt {
-            job_id: icn_common::Cid::new_v1_dummy(0x55, 0x12, b"r"),
-            executor_did: did.clone(),
-            result_cid: icn_common::Cid::new_v1_dummy(0x55, 0x12, b"r"),
-            cpu_ms: 1000,
-            success: true,
-            sig: icn_identity::SignatureBytes(vec![]),
-        };
-        store.record_execution(&receipt.executor_did, receipt.success, receipt.cpu_ms);
-        assert_eq!(store.get_reputation(&did), 2);
-        drop(store);
-        let reopened = SledReputationStore::new(path).unwrap();
-        assert_eq!(reopened.get_reputation(&did), 2);
     }
 }
